Create and map a ring
RT-005, phase two: rt_ring_setup creates a ring with SINGLE_ISSUER |
DEFER_TASKRUN | NO_SQARRAY, maps it, and derives every pointer the runtime
needs. Verified before commit against a temporary probe, which reported:

    setup 0 fd 3 sqmask 7 cqmask 15 ringlen 320

sqmask 7 is the 8 entries asked for; cqmask 15 is the kernel's default
doubling of the CQ; ringlen 320 is cq_off.cqes (64) plus sixteen 16-byte CQEs.

The ring length is computed from cq_off.cqes, not from sq_off.array. liburing's
formula uses the latter, and under NO_SQARRAY the kernel never assigns that
field (io_uring.c:2941), so it stays zero and the arithmetic yields a mapping
far too small — mmap succeeds and the fault arrives much later on a CQE read,
looking exactly like a missing barrier.

Deriving the pointers needs seven reinterpretations of a byte pointer, each of
which -Wcast-align objects to. Casting through void * would silence it by
hiding the cast; __builtin_assume_aligned instead states the claim, and
-fsanitize=alignment compiles it into a runtime check (measured: two extra
instructions). The claim holds because mmap returns page-aligned memory and the
kernel lays these fields out naturally aligned within it — but it is now
checked rather than asserted, locally at least. The container build carries no
UBSan until RT-007 supplies the __ubsan_handle_* symbols; there are five of
them, so that is a small ticket away from being checked in the shipped binary
too.

Two kinds of offset, treated differently: head and tail change under us and
stay pointers, while the masks never change after setup and are read once and
cached as values. Entry counts are mask + 1 and are not stored at all.

Build: -Weverything, minus seven families that fight this codebase rather than
check it — C23 and C++ compat noise, a C89 declaration rule, struct padding
(rt_ctx's static_assert covers where it matters), missing-noreturn (which would
demand a false contract on every trap stub), unsafe-buffer-usage (it wants
std::span), and nix's cc-wrapper injecting -mmacos-version-min. It immediately
found two real defects: a silent long-to-int narrowing, and rt_main having no
prototype anywhere despite being called from assembly.

Switching the local compiler to clang 22 also made the UBSan probe pass for the
first time — Apple clang has no minimal runtime, so -fsanitize=undefined has
been silently absent from every local build until now.

+++ src/syscall.h
@@ -210,8 +210,8 @@ static inline long raw_write(int fd, const void *buf, size_t len) {
  * Direct syscall, not a ring op: mmap has no io_uring opcode, which is
  * why invariant 1 lists it as permitted.
  */
-static inline long sys_mmap(void *addr, size_t len, int prot, int flags,
-                            int fd, unsigned long off) {
+static inline long sys_mmap(void *addr, size_t len, int prot, int flags, int fd,
+                            unsigned long off) {
   return sys6(__NR_mmap, (long)(uintptr_t)addr, (long)len, prot, flags, fd,
               (long)off);
 }
@@ -221,8 +221,8 @@ static inline long sys_mmap(void *addr, size_t len, int prot, int flags,
  * Three arguments, so sys3 covers it. Used to open the usable part of a
  * task stack after mapping the whole region PROT_NONE.
  */
-static inline long sys_mprotect(void *addr, size_t len, int prot) {
-  return sys3(__NR_mprotect, (long)(uintptr_t)addr, (long)len, prot);
+static inline int sys_mprotect(void *addr, size_t len, int prot) {
+  return (int)sys3(__NR_mprotect, (long)(uintptr_t)addr, (long)len, prot);
 }
 
 /* Incomplete type: enough to declare a pointer parameter, and it keeps a
@@ -243,22 +243,22 @@ struct io_uring_params;
  * "blind" path that needs no ring at all (register.c:1031), which is how the
  * capability probe runs before setup has been called.
  */
-static inline long sys_io_uring_setup(unsigned entries,
-                                      struct io_uring_params *p) {
-  return sys2(__NR_io_uring_setup, entries, (long)(uintptr_t)p);
+static inline int sys_io_uring_setup(unsigned entries,
+                                     struct io_uring_params *p) {
+  return (int)sys2(__NR_io_uring_setup, entries, (long)(uintptr_t)p);
 }
 
-static inline long sys_io_uring_enter(int fd, unsigned to_submit,
-                                      unsigned min_complete, unsigned flags,
-                                      const void *arg, size_t argsz) {
-  return sys6(__NR_io_uring_enter, fd, to_submit, min_complete, flags,
-              (long)(uintptr_t)arg, (long)argsz);
+static inline int sys_io_uring_enter(int fd, unsigned to_submit,
+                                     unsigned min_complete, unsigned flags,
+                                     const void *arg, size_t argsz) {
+  return (int)sys6(__NR_io_uring_enter, fd, to_submit, min_complete, flags,
+                   (long)(uintptr_t)arg, (long)argsz);
 }
 
-static inline long sys_io_uring_register(int fd, unsigned opcode, void *arg,
-                                         unsigned nr_args) {
-  return sys4(__NR_io_uring_register, fd, opcode, (long)(uintptr_t)arg,
-              nr_args);
+static inline int sys_io_uring_register(int fd, unsigned opcode, void *arg,
+                                        unsigned nr_args) {
+  return (int)sys4(__NR_io_uring_register, fd, opcode, (long)(uintptr_t)arg,
+                   nr_args);
 }
 
 #endif /* RT_SYSCALL_H */
